--- services/company-service/src/main/java/com/sixlab/logistics/company_service/BaseEntity.java
@@ -1,46 +1,0 @@
-package com.sixlab.logistics.company_service;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-    // 생성, 수정 관련 필드는 JPA Auditing으로 자동 주입
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    @CreatedBy
-    @Column(updatable = false)
-    private String createdBy;
-
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-
-    @LastModifiedBy
-    @Column
-    private String updatedBy;
-
-    // 삭제 관련 필드는 각 엔티티 삭제시 직접 주입 해야함!
-    @Column
-    private LocalDateTime deletedAt;
-
-    @Column
-    private String deletedBy;
-
-    public void delete(String deletedBy) {
-        this.deletedAt = LocalDateTime.now();
-        this.deletedBy = deletedBy;
-    }
-}
