--- services/company-service/src/main/java/com/sixlab/logistics/company_service/JpaConfig.java
@@ -1,15 +1,0 @@
-//package com.sixlab.logistics.company_service;
-//
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.data.domain.AuditorAware;
-//import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-//
-//@Configuration
-//@EnableJpaAuditing(auditorAwareRef = "auditorAware")
-//public class JpaConfig {
-//    @Bean
-//    public AuditorAware<String> auditorAware() {
-//        return new AuditorAwareImpl();
-//    }
-//}
